% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar-compressed-matrix.R
\docType{data}
\name{GrammarCompressedMatrix}
\alias{GrammarCompressedMatrix}
\title{Grammar Compressed Matrix Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GrammarCompressedMatrix
}
\description{
Grammar Compressed Matrix Class
}
\keyword{datasets}

