% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpls.R
\name{cpls}
\alias{cpls}
\title{Compression-based Partial Least Squares Regression (cPLS)}
\usage{
cpls(formula, data = NULL, ncomp = 10, re_pair_method = c("freq", "naive",
  "lossless", "lossy"), verbose = TRUE)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{ncomp}{ncomp}

\item{re_pair_method}{re_pair_method}

\item{verbose}{verbose}
}
\description{
Compression-based Partial Least Squares Regression (cPLS)
}

